progress on routine, need to implement countdown to death

+++ includes/philo.h
@@ -47,10 +47,10 @@ typedef struct s_philo
 	bool			is_alive;
 	pthread_t		thread;
 	t_mutex_list	*mutex_list;
-	size_t			time_left_death;
-	size_t			time_left_eat;
-	size_t			time_left_sleep;
-	size_t			loop_left;
+	int				time_left_death;
+	int				time_left_eat;
+	int				time_left_sleep;
+	int				loop_left;
 	t_fork			*left_fork;
 	t_fork			*right_fork;
 	int				initial_time;
